refactor: useContainer, useWorkBlocks 분리
