+++ frontend/src/pages/homePage/contexts/WorkBlocksContext.tsx
@@ -8,8 +8,6 @@ export interface WorkBlocksContextType {
   updateWorkBlocks: (updatedBlocks: WorkBlockType[]) => void;
   removeWorkBlock: (id: number | string) => void;
   containerRef: RefObject<HTMLDivElement | null>;
-  scrollOffset: number;
-  setScrollOffset: (offset: number) => void;
   selectedRecommendedWork: RecommendedWorksResponse | null;
   setSelectedRecommendedWork: (work: RecommendedWorksResponse | null) => void;
 }
@@ -20,8 +18,6 @@ const WorkBlocksContext = createContext<WorkBlocksContextType>({
   updateWorkBlocks: () => {},
   removeWorkBlock: () => {},
   containerRef: { current: null },
-  scrollOffset: 0,
-  setScrollOffset: () => {},
   selectedRecommendedWork: null,
   setSelectedRecommendedWork: () => {},
 });
