+++ frontend/src/pages/homePage/hooks/useCreateWorkBlock.ts
@@ -5,11 +5,12 @@ import type {
   MyCropResponse,
   RecommendedWorksResponse,
 } from '@/types/openapiGenerator';
-import useWorkBlocks from '@/contexts/useWorkBlocks';
 import calculateTimeToPosition from '../utils/calculateTimeToPosition';
 import { getYCoordinate } from '@/constants/workTimeCoordinate';
 import { findCollisionFreePosition } from '@/utils/collisionUtils';
 import updateBlockWorkTime from '@/pages/homePage/utils/updateBlockWorkTime';
+import useContainer from '@/pages/homePage/contexts/useContainer';
+import useWorkBlocks from '@/pages/homePage/contexts/useWorkBlocks';
 
 interface UseCreateWorkBlockReturn {
   handleCreateWork: (
@@ -19,8 +20,8 @@ interface UseCreateWorkBlockReturn {
 }
 
 export const useCreateWorkBlock = (): UseCreateWorkBlockReturn => {
-  const { addWorkBlock, workBlocks, containerRef, scrollOffset } =
-    useWorkBlocks();
+  const { containerRef, scrollOffset } = useContainer();
+  const { addWorkBlock, workBlocks } = useWorkBlocks();
 
   const handleCreateWork = useCallback(
     (work: RecommendedWorksResponse, selectedCrop: MyCropResponse) => {
